requirements.txt
beautifulsoup4==4.12.2
Flask==2.3.2
psycopg2==2.9.9
python-dotenv==1.0.0
Requests==2.31.0
